pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "tiled_ingestor"
description = "Process that listens on a queue for items to add to tiled"
readme = { file = "README.md", content-type = "text/markdown" }

requires-python = ">=3.11"



classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11"
]

dependencies = [
   "pika",
   "stomp.py",
   "python-dotenv"
]

dynamic = ["version"]


[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pre-commit",
    "flake8"
]

file_watcher = [
    "pandas",
    "python-dotenv==1.0.0",
    "h5py==3.9.0",
    "fabio==2023.6.0",
    "tiled[all]==0.1.0a116",
]

[tool.hatch]
version.source = "vcs"
build.hooks.vcs.version-file = "src/_version.py"

[tool.hatch.metadata]
allow-direct-references = true